+++ backend/database/migrations/2026_07_30_000002_create_ticket_internal_notes_table.php
@@ -8,16 +8,16 @@ return new class extends Migration
 {
     public function up(): void
     {
-        // If a previous failed migration left a partial table, recreate it cleanly.
         if (Schema::hasTable('ticket_internal_notes')) {
             Schema::dropIfExists('ticket_internal_notes');
         }
 
         Schema::create('ticket_internal_notes', function (Blueprint $table) {
             $table->id();
-            // Match existing legacy schema where user and ticket IDs are INT.
-            $table->integer('ticket_id');
-            $table->integer('user_id');
+
+            $table->unsignedBigInteger('ticket_id');
+            $table->unsignedBigInteger('user_id');
+
             $table->text('note');
             $table->timestamps();
